admin customer - bug fix/enhancement
There was a bug on the API level that caused customer records requested
from any non-zero partition to come back blank. This was fixed.

The UI does require that the customer record and domain in focus be on
the same partition. The 'list' mode was updated to reflect this
requirement.