admin - enhancement: send email
Added an email tool for sending. supports a 'list' type which will fetch
the system messages that match (ex: CUSTOMER or ORDER) and show them in
a chooser along with a 'blank' option. Selecting a predefined source
will populate the body and allow customization and even let the changes
be saved back to the original message. Message list is partition
specific. Currently used in customer manager.