search kenburns/masonry, mushpot not destroyed
Added kenburns and masonry init to store_search so they are ran at the
correct time (content added to page after oncompletes here).
Renderformat to retrieve images from elastic differently for ken bunrs
init. (to use the canvas will have to set up ken burns function it self
to get image data differently).
Set up if in oncompletes for mushpot (on hompage so far) so it is still
there when the page is revisited.