prod page not avail purchse pid with dash/version
Some product id's had dashes in them, but the DOM wasn't showing them
with dashes. So those that weren't available for purchase with dashes
were not getting the icon on the add to cart button removed.
Split the pid on the dash and added it back together to add the
displayNone class in that situation.