contact links set to reset pid entry
If the ask a question button is used on the product page to get to the
contact form it puts the pid in one of the fields for reference. If the
form isn't submitted and the user returns via one of the other contact
us page links in the app it will reset the pid field to the default
placeholder.